Major refactoring

+++ src/soot/dexpler/instructions/BinopLitInstruction.java
@@ -30,7 +30,6 @@ import org.jf.dexlib2.iface.instruction.NarrowLiteralInstruction;
 import org.jf.dexlib2.iface.instruction.TwoRegisterInstruction;
 import org.jf.dexlib2.iface.instruction.formats.Instruction22b;
 import org.jf.dexlib2.iface.instruction.formats.Instruction22s;
-
 import soot.Local;
 import soot.Value;
 import soot.dexpler.Debug;
@@ -48,14 +47,14 @@ public class BinopLitInstruction extends TaggedInstruction {
 
     Value expr = null;
     AssignStmt assign = null;
-  
-    public BinopLitInstruction (Instruction instruction, int codeAdress) {
+
+    public BinopLitInstruction(Instruction instruction, int codeAdress) {
         super(instruction, codeAdress);
     }
 
-    public void jimplify (DexBody body) {
-        if(!(instruction instanceof Instruction22s) && !(instruction instanceof Instruction22b))
-            throw new IllegalArgumentException("Expected Instruction22s or Instruction22b but got: "+instruction.getClass());
+    public void jimplify(DexBody body) {
+        if (!(instruction instanceof Instruction22s) && !(instruction instanceof Instruction22b))
+            throw new IllegalArgumentException("Expected Instruction22s or Instruction22b but got: " + instruction.getClass());
 
         NarrowLiteralInstruction binOpLitInstr = (NarrowLiteralInstruction) this.instruction;
         int dest = ((TwoRegisterInstruction) instruction).getRegisterA();
@@ -63,7 +62,7 @@ public class BinopLitInstruction extends TaggedInstruction {
 
         Local source1 = body.getRegisterLocal(source);
 
-        IntConstant constant = IntConstant.v((int)binOpLitInstr.getNarrowLiteral());
+        IntConstant constant = IntConstant.v(binOpLitInstr.getNarrowLiteral());
 
         expr = getExpression(source1, constant);
 
@@ -73,87 +72,87 @@ public class BinopLitInstruction extends TaggedInstruction {
         setUnit(assign);
         addTags(assign);
         body.add(assign);
-        
+
         if (IDalvikTyper.ENABLE_DVKTYPER) {
-			Debug.printDbg(IDalvikTyper.DEBUG, "constraint: "+ assign);
-            int op = (int)instruction.getOpcode().value;
-          if (op >= 0xd8) {
-            op -= 0xd8;
-          } else {
-            op -= 0xd0;
-          }
-          BinopExpr bexpr = (BinopExpr)expr;
-          //body.dvkTyper.setType((op == 1) ? bexpr.getOp2Box() : bexpr.getOp1Box(), op1BinType[op]);
-          DalvikTyper.v().setType(((JAssignStmt)assign).leftBox, op1BinType[op], false);
+            Debug.printDbg(IDalvikTyper.DEBUG, "constraint: " + assign);
+            int op = (int) instruction.getOpcode().value;
+            if (op >= 0xd8) {
+                op -= 0xd8;
+            } else {
+                op -= 0xd0;
+            }
+            BinopExpr bexpr = (BinopExpr) expr;
+            //body.dvkTyper.setType((op == 1) ? bexpr.getOp2Box() : bexpr.getOp1Box(), op1BinType[op]);
+            DalvikTyper.v().setType(((JAssignStmt) assign).leftBox, op1BinType[op], false);
         }
     }
 
     @SuppressWarnings("fallthrough")
     private Value getExpression(Local source1, Value source2) {
-      Opcode opcode = instruction.getOpcode();
-        switch(opcode) {
-        case ADD_INT_LIT16:
-          setTag (new IntOpTag());
-        case ADD_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newAddExpr(source1, source2);
+        Opcode opcode = instruction.getOpcode();
+        switch (opcode) {
+            case ADD_INT_LIT16:
+                setTag(new IntOpTag());
+            case ADD_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newAddExpr(source1, source2);
 
-        case RSUB_INT:
-          setTag (new IntOpTag());
-        case RSUB_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newSubExpr(source2, source1);
+            case RSUB_INT:
+                setTag(new IntOpTag());
+            case RSUB_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newSubExpr(source2, source1);
 
-        case MUL_INT_LIT16:
-          setTag (new IntOpTag());
-        case MUL_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newMulExpr(source1, source2);
+            case MUL_INT_LIT16:
+                setTag(new IntOpTag());
+            case MUL_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newMulExpr(source1, source2);
 
-        case DIV_INT_LIT16:
-          setTag (new IntOpTag());
-        case DIV_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newDivExpr(source1, source2);
+            case DIV_INT_LIT16:
+                setTag(new IntOpTag());
+            case DIV_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newDivExpr(source1, source2);
 
-        case REM_INT_LIT16:
-          setTag (new IntOpTag());
-        case REM_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newRemExpr(source1, source2);
+            case REM_INT_LIT16:
+                setTag(new IntOpTag());
+            case REM_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newRemExpr(source1, source2);
 
-        case AND_INT_LIT8:
-          setTag (new IntOpTag());
-        case AND_INT_LIT16:
-          setTag (new IntOpTag());
-            return Jimple.v().newAndExpr(source1, source2);
+            case AND_INT_LIT8:
+                setTag(new IntOpTag());
+            case AND_INT_LIT16:
+                setTag(new IntOpTag());
+                return Jimple.v().newAndExpr(source1, source2);
 
-        case OR_INT_LIT16:
-          setTag (new IntOpTag());
-        case OR_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newOrExpr(source1, source2);
+            case OR_INT_LIT16:
+                setTag(new IntOpTag());
+            case OR_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newOrExpr(source1, source2);
 
-        case XOR_INT_LIT16:
-          setTag (new IntOpTag());
-        case XOR_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newXorExpr(source1, source2);
+            case XOR_INT_LIT16:
+                setTag(new IntOpTag());
+            case XOR_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newXorExpr(source1, source2);
 
-        case SHL_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newShlExpr(source1, source2);
+            case SHL_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newShlExpr(source1, source2);
 
-        case SHR_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newShrExpr(source1, source2);
+            case SHR_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newShrExpr(source1, source2);
 
-        case USHR_INT_LIT8:
-          setTag (new IntOpTag());
-            return Jimple.v().newUshrExpr(source1, source2);
+            case USHR_INT_LIT8:
+                setTag(new IntOpTag());
+                return Jimple.v().newUshrExpr(source1, source2);
 
-        default :
-            throw new RuntimeException("Invalid Opcode: " + opcode);
+            default:
+                throw new RuntimeException("Invalid Opcode: " + opcode);
         }
     }
